requirements.txt
# This file represents the needs for the deployer script to function, while the
# dev-requirements.txt file represents the needs in this repo in general.
#

# ruamel.yaml is used to read and write .yaml files.
ruamel.yaml==0.17.*

# jsonschema is used for validating cluster.yaml configurations
jsonschema==4.*

# rich, py-markdown-table, tabulate are used for pretty printing outputs that would otherwise
# be difficult to parse by a human
rich==13.*
py-markdown-table==0.3.*
tabulate==0.9.*

# jhub_client, pytest, and pytest-asyncio are used for our health checks
jhub-client==0.1.*
pytest
pytest-asyncio

# Used to generate templates
jinja2==3.*

# Used for the debug CLI
typer==0.9.*
escapism==1.*

# Used for GCP billing management
pandas
pandera
prometheus_pandas
google-cloud-bigquery==3.10.*
google-cloud-bigquery[pandas]==3.10.*
gspread==5.9.*

# requests is used by deployer/cilogon_app.py
requests==2.*

# auth0 is used to communicate with Auth0's REST API that we support as a deployer sub-cmd
auth0-python==3.*
